src/pyproject.toml
[project]
name = "TEBTRMNL"
version = "0.1.0"
description = "A Telegram bot for reading documents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "python-telegram-bot>=22.3",
    "requests>=2.32.4",
]
